Refactor Pathlike implementation into Turn
