+++ lib/chukinas/geometry/turn.ex
@@ -1,5 +1,5 @@
-alias Chukinas.Geometry.{PathLike, Pose}
-alias Chukinas.Geometry.Path.Turn
+alias Chukinas.Geometry.{Pose, Position, PathLike, Point, Rect, Straight, Turn}
+
 defmodule Turn do
 
   use TypedStruct
@@ -24,6 +24,29 @@ defmodule Turn do
   end
   def new(x, y, angle, len, angle) do
     new Pose.new(x, y, angle), len, angle
+  end
+
+  # *** *******************************
+  # *** GETTERS
+
+  def start_pose(path), do: path.pose
+  def end_pose(path) do
+    right_angle = (90 * get_sign path.angle)
+    path.pose
+    |> Pose.rotate(right_angle)
+    |> Straight.new(path.radius)
+    |> PathLike.pose_end()
+    |> Pose.rotate(180 + path.angle)
+    |> Straight.new(path.radius)
+    |> PathLike.pose_end()
+    |> Pose.rotate(right_angle)
+  end
+  def bounding_rect(path) do
+    {x_start, y_start} = path |> start_pose() |> Position.to_tuple()
+    {x_end, y_end} = path |> end_pose() |> Position.to_tuple()
+    {xmin, xmax} = Enum.min_max([x_start, x_end])
+    {ymin, ymax} = Enum.min_max([y_start, y_end])
+    Rect.new(Point.new(xmin, ymin), Point.new(xmax, ymax))
   end
 
   # *** *******************************
@@ -53,7 +76,21 @@ defmodule Turn do
   # *** *******************************
   # *** PRIVATE
 
+  defp get_sign(number) when number > 0, do: 1
+  defp get_sign(_number), do: -1
+
   defp radius(length, angle) do
     (length * 360) / (2 * :math.pi() * abs(angle))
   end
+
+  # *** *******************************
+  # *** IMPLEMENTATIONS
+
+  defimpl PathLike do
+    def pose_start(path), do: path.pose
+    def pose_end(path), do: Turn.end_pose(path)
+    def len(path), do: path.length
+    def get_bounding_rect(path), do: Turn.bounding_rect(path)
+  end
 end
+
